Add a workaround for size 0 depths in Dask
Appears that Dask has a bug when size 0 depths are present.
Unfortunately we can't simply skip adding halos in this case as we may
need halos on some dimensions and not others. While we could special
case depths that are 0 on every dimension, this just complicates the
code on our end. So we add an easy workaround that adds a 1 element halo
to all borders no matter what. As this seems to workaround the problem
sufficiently, this should be fine until Dask solves this properly.