Improve enforcement of real values for the origin
Make sure any single number is converted into an array. This include
types like complex that are definitely not supported. Once we have an
array, enforce that the type is definitely real. By this we should
capture a few other strange cases like a complex origin array or
similar. To ensure we are raising this error correctly, also include
some tests with unacceptable types for the origin.