Require origin to be of integral type
As only some of the filters handle an origin of floating type by
effectively converting it to an origin of integral type and other
filters simply disallow it, it seems like it would just make more sense
to simply disallow an origin of floating type instead of handling it.
After all this should just be clearer for the end user that they need to
specify a particular index. Allowing a floating type just results in
confusion about how this is handled. If the user has to handle the
floating type themselves, then there is no lack of clarity as to how
this was done. We do this regardless of filter even if the underlying
function allows it. Further we add some tests to verify that floating
type origins are disallowed. Verified that all existing tests of
floating type origin resulted in failures now as well.