Drop _get_normed_args
Instead of having a specialty function for handling size, footprint, and
origin to provide cleaned up versions of those parameters and some
additional ones for halos, simply make the calls for each parameter
individually. This results in a small amount of code duplication, but
seems ok as opposed to get a bunch of different parameter sanitizing
functions.