Exit testing with pytest exit code
Previously there was no non-zero exit code on test failure as pytest
merely returns its exit code and we weren't handling it. To fix this
pass the exit code to `sys.exit`. This way it is properly propagated up
the chain and running the tests from `setup.py` should set a non-zero
exit code.