Vendor fftfreq from dask for compatibility
This currently exists in an unreleased version of Dask. Also, once
released, it still won't be present in Python 3.4 packages from
`defaults` or `conda-forge`. So we will still want to keep it vendored
until we drop Python 3.4 and bump our Dask requirement. Until then we
can keep this vendored. After all it is our contribution.