Workaround Dask Array concatenate bug for fftfreq
Appears that old Dask Array versions encounter a bug with `concatenate`
when used with empty arrays. This adds a hacky workaround for those
cases to detect if an empty array will occur on the right side (the only
place it will occur) before trying to concatenate them. If the right
side is found to be empty, then skip concatenating and just keep the
left end.