Add a function for computing the frequency grid
In order to compute things in Fourier space, we need to have some notion
of what the frequencies are for each point in the array. This adds a
function, which provides this grid of frequencies for each dimension
given the shape of the data, type, and chunk size. Once generated this
Dask Array can be consumed by other functions to perform computation in
Fourier space.