Parameterize all tests based on Fourier filter
Looking at the functions in SciPy ndimage's Fourier filters, they all
appear to take some argument after the array that is either a single
value or 1-D sequence of values. In either case, if they are all zero it
always results in an identity. If the type constraints aren't met, then
the same sorts of issues are raised. In all cases we want to make sure
they match the equivalent SciPy function. So we can parameterize over
filter function name without loss of generality.