Join negative and zero valued tests
Zero values for filter sizes and shifts are a bit of a strange thing.
Stranger still are negative values, which breakdown on many filters.
Surprisingly they work well for the Gaussian filter in SciPy (less
surprisingly for our implementation). However these are really odd to
consider for the uniform filter. So go ahead and move these out of the
standard filter tests and join them with the negative size and shift
tests. That way these can be opted into if they make sense.