Break up the identity and type tests
For the identity test we don't really need to concern ourselves with
types if we always use complex as the result will always be complex. So
just do that. As for the type test, we want to be able to test things
that don't have an identity operation. So simply pick some reasonable
shift, size, or similar that will work for this sort of test and merely
permute types to make sure our implementations behave like SciPy's.