Add a Dask Array sinc ufunc for compatibility
Future versions of Dask Array will have `sinc` as a ufunc. However there
are no current releases that have `sinc` yet. Further the Python 3.4
packages of Dask Array are not likely to see a new release on `defaults`
or `conda-forge`. So we will need carry this compatibility function
until we drop Python 3.4.