Relax sequence type constraint on `s`
Instead of requiring that `s` be an instance of `collections.Sequence`,
simply format it as a NumPy array if it is a single number. Allow Dask
Arrays to be passed through unchanged type normalization unchanged. In
addition, adjust type and shape checks to verify that `s` is 1-D and the
right length (1-D was satisfied by the `collections.Sequence` check).
Also update the type check to make use of the `dtype` attribute, which
should simplify things.