Convert s to float in type normalization
In all cases, we want `s` to be real in the end. This is because it will
always end up multiplying real values from the frequency grid. So we
might as well handle the type conversion beforehand. Here we make sure
that if `s` is of an integral type, that we convert it to the same real
type used by the input array. This ensures that we won't end up with any
less precision than the input array.