Specify the frequency grid type to match s
As the (angular) frequency grid will always end up multiplying `s` first
to construct the filter, go ahead and make sure it has the same type as
`s` does as well. Given that we know `s` will always have a real type
thanks to argument normalization, this should work without issues.

There is no need to handle type conversion to the input array explicitly
when computing the frequency grid. Even if there is a mismatch in types,
this will be resolved automatically when multiplying the Fourier filter
and the input array.