Test precision is matched in type test
If the type does not need to be upcast, make sure the precision is
matched. In other words, ensure single precision results are gotten when
single precision data is provided.

However, if the type does need to be upcast, ensure that we just get
double precision as we would expect. This is no different than before.