Import dask to avoid a sporadic threading issue
Seems there is an occasional sporadic threading issue that comes up with
Dask when running the test suite through `python setup.py test` on
Python 2. While it doesn't fail the test suite, it does add some noise
about threads shutting down and having nothing to run. Not entirely sure
why this happens or why this helps, but have added these imports as they
seem to suppress the issue. Also added a simple assert statement for the
dask import to make sure it is not flagged as unused.