Drop setting `name` in `map_overlap`
Dask apparently requires the `name` passed to `map_blocks`/`map_overlap`
to be unique for each computation. So the `name` typically has a hash of
the arguments tacked onto it. However when the `name` is specified
externally this hashing doesn't occur and it is expected the user will
do their own hashing. As we weren't handling the hashing previously, it
seems the most recent computation would take the value for the hash
regardless of what array it was computed on or any other arguments. For
now simply drop passing `name` so that hashing will be done for us. We
can revisit setting `name` in the future if we need to.