Require slicerator
`slicerator` is a dependency of `pims`. However for some reason it
doesn't seem to get pulled in correctly with the `conda-forge` package
that we are testing against. So we add an explicit requirement in our
environment files for the CIs and Docs. This should hopefully rectify
that issue and ensure `slicerator` is installed correctly with `conda`
before proceeding.