Require slicerator from pip
As explicitly adding `slicerator` to the environment file for `conda` to
install does not seem to avoid `pip` installing `slicerator` later when
we install our package with `pip`, simply switch to using `pip` to
install `slicerator` in our environment file. That way `pip` should know
`slicerator` is already installed as it has done the installation
itself.