Require slicerator from conda and pip
To make sure that the versions match up between the `conda` installed
`slicerator` and the `pip` installed `slicerator`. Pin the `conda`
version and the `pip` version to the same version. That way even though
`pip` is overwriting the `conda` installed version of the package, we
can be guaranteed that the same compatibility can be expected from both.