Test loading TIFFs with imread
Try writing out a variety of temporary test TIFF files to disk with
scikit-image's tifffile. Construct a Dask Array from the TIFF file using
our `imread` function. Then compare the test data that we wrote to the
TIFF file to that read from our Dask Array. This all to validate we are
able to read the TIFF file through Dask correctly.