Try to import tifffile if possible
First try to import `tifffile` directly. As this is a dependency of
`pims` in most cases this should work.  However some older versions of
`pims` (particularly for Python 3.4) pull in `scikit-image` instead for
some reason. So if `tifffile` can't be imported directly, try importing
it from `scikit-image`'s `external` package.