Monkey patch pims to use the available tifffile
This seems to occur on Python 3.4 as the pims 0.3.3 package on
conda-forge pulls in scikit-image instead of tifffile, but seems to not
be able to use it. To fix that we import any available tifffile and
monkey patch pims only if it didn't find any tifffile. However there is
a chance we failed to find tifffile as well. In that case this amounts
to a no-op.