Ignore test coverage when scikit-image is missing
Even though it is extremely unlikely that both `tifffile` and
`scikit-image` will not be present, we would still like to gracefully
handle this case. Instead of dropping the exception handling here
because it is  technically not covered in our testing, simply note to
coverage that it should not bother with this case.