Treat `nframes` equal to `-1` as one big chunk
If the user provides `nframes` equal to `-1`, simply replace `nframes`
with the number of frames in the image. In other words, the entire image
will be loaded into memory for computation. This may be handy in image
data that is spread out amongst many small image files. For cases like
this, it may be completely reasonable to read the whole image into
memory as IO may be more expensive than the memory consumed. Include
tests to make sure that this value of `nframes` behaves as expected both
in terms of being able to load the data and provide the expected number
of chunks, namely `1`.