Test Fourier filters with a chunked `s`
Just to make sure that chunking doesn't get messed up, add a test with a
chunked `s`. Really this should effect anything as the Fourier filter is
normally reduced along `s`. So any chunking of `s` should already be a
non-issue by the time the Fourier filter is applied on the data. Still
this is a good sanity check.