Add a test for different shapes and chunks
Includes a test that tries different combinations of shapes and chunks
with the Fourier filters. The purpose of this test is both to ensure the
chunking remains unchanged and to ensure that the results remain
accurate. Though as these operations are element-wise, chunking is
unlikely to be a cause of inaccuracies unlike with other filters, which
require an overlap.