Vendor argwhere from our Dask contribution
Basically verbatim the same as what we submitted upstream to Dask for
the `argwhere` function. Though in Dask 0.13.0 this is not a lazy
implementation due to `compress` being non-lazy. Support for unknown
dimensions in Dask 0.13.0 is in its nascent stages. So it may be
difficult to get a lazy implementation that works ok there, but will
give that a try. Fortunately newer versions of Dask (not yet released)
should have this builtin.