Restrict some tests to certain Dask versions
Not all of the tests for the vendored compatibility functions can be run
with all versions of Dask. This is due to some missing features and/or
bugs in previous versions of Dask. As these tests were never designed to
run on such an old version, we simply skip some old versions of Dask for
some tests if they don't work. There is no point in trying to "fix" them
as the code was designed with a newer version of Dask in mind. Also, at
some point, we intend to drop the compatibility functions once a newer
version of Dask is required.