Include tests for _assert_eq_nan
Just provides some basic tests to make sure it can compare to arrays
that are the same and find that they are equal. Also checks that when
arrays shouldn't be equal that the function raises an `AssertionError`.
Include arrays with and without NaN. Also includes different types of
arrays to make sure it behaves ok.