Implement center_of_mass using Dask Arrays
Provides a basic implementation of `center_of_mass` akin to SciPy
ndimage's function except that it uses Dask Arrays instead of NumPy
arrays. Also as SciPy's implementation seems to support multiple
indices, we try to do the same here. SciPy uses a variety of lists and
arrays in their result. We do not bother to do this here for a few
reasons. First it makes it harder to manipulate the data. Second it
makes evaluation of the result more cumbersome. Third the structure is
not of particular value.