Include tests for  center_of_mass
Compares the result between SciPy's implementation and Dask's
implementation. Some of the tests include just providing the array,
including a label image, and including a label image and label indices.
For label indices a variety of formats are tried including scalar, 1-D
array, and various N-D arrays. These are important to test as SciPy
allows all of them and formats the output differently depending on how
the user provides them. Also we test the case where the label index
provided is not present (results in `nan`s for the center of mass).