Vendor/refactor lazy implementation of compress
This is needed for some other use cases that `dask-ndmeasure` is trying
to cover. Filtering out labels that are not being used is essential in
these cases and replacing them with 0 simply won't work. Though it
should be able to speed up cases where we did replace values with 0.

While `compress` will be included in newer (not yet released) versions
of Dask, it is not present in any released version and it is especially
not in Dask 0.13.0, which is needed for Python 3.4 support. So for now
we include a simple implementation of `compress` that is based on code
in the vendored `argwhere` with a few tweaks and validation checks.