Move index transpose into center_of_mass
The index transposing behavior (i.e. the selected labels) appears to be
unique to the `center_of_mass` computation or at least it is not present
in `sum`. So this moves that tweak into the `center_of_mass`
computation. That way it doesn't contaminate other computations that do
not need these. Further it allows the argument normalization function to
be reused repeatedly without ill effect.