Add labeled_comprehension
Provides an implementation of `labeled_comprehension` for use with Dask
Arrays. Mimics the implementation from SciPy ndimage. Should be possible
to rewrite everything in terms of `labeled_comprehension` and make use
of it in implementations of other computations that have not yet been
added. Also this should make it easy to support all other manner of
computations over label images that are not already (or will not be)
covered by the API.