Fix test label array construction
Was using Dask Array's in the construction of the label array, which
works on old versions of Dask, but not newer versions. As we really want
to use NumPy arrays in the construction of all test data and that will
fix the problem, go ahead and switch the label array construction to use
NumPy arrays exclusively.