Fix testing of empty arrays with compress
Before the test data generated was being sliced if `axis` was specified.
However this makes no sense if the test data has an empty dimension as
it may mean slicing along the 0 length axis. Even if slicing doesn't
occur along the 0 length axis, we already know the result will be a
length 0 1-D array. We can get to this result much faster by simply
flattening the array instead of slicing. Hence we use this strategy to
fix the tests with newer versions of Dask.