Handle `labeled_comprehension` comparison directly
As Dask Arrays using delayed appear to be invalid according to its own
array comparison test (particularly in Dask 0.14.0+), simply handle the
comparison of the two results ourselves. This way we can be sure that
the two results are treated correctly regardless of which version of
Dask is in use.