Rewrite of maximum that uses nanmax
This basically is a clever trick to avoid using `labeled_comprehension`
in the computation of `maximum`. It works by converting the data to
`float64` and placing `nan` over all unneeded values. This way `nanmax`
can be run and all of the `nan` values will be ignored (unless there are
only `nan` values). Using info about which labels could be computed,
`where` is run over the final result to replace those values that could
not be computed with `0`. The result is then cast back to its original
type. This change up should allow the computation to proceed faster than
it would by going through `delayed` in `labeled_comprehension`.