Add a 3-D test case to the mix
It is pretty easy to mess things up, but still get the right result in
the 2-D test case (e.g. indexing by 1 or -1 is the same in 2-D). Having
a 3-D test case breaks symmetry for some cases, which makes it a bit
easier to check if the code is doing the right thing.