Name all arguments in _labeled_comprehension_func
As the user specified function, `func`, needs to take a particular
argument structure that is known, there is no need to leave this so
vague in `_labeled_comprehension_func`. Doing this does mean there is an
additional check each time the function is run. However having the
arguments specified will help us remove things like `compute`, which
should speed up the computation on the Dask end.