labeled_comprehension: Get position before `for`
Handle the check as to whether to compute or include the position before
entering the label comprehension `for`-loop. This not only avoids
repeating this check inside the `for`-loop, but it avoids constructing
the raveled indices if the won't be used. Should be more efficient when
the position is not needed in the user function.