Fused delayed steps in labeled_comprehension
Instead of calling `dask.delayed` on the function and storing it before
converting it to a Dask Array (and storing it again), simply pass the
result of `dask.delayed` directly into `dask.array.from_delayed` before
storing it in the NumPy object array. This cuts down on the
retrieval/stores to NumPy object array and should help streamline things
a bit. May want to consider refactoring this into `_utils` a bit to
cutdown on the visual clutter.