Construct index ranges once in comprehensions
The `range`s were being constructed each time the `for`-loop went
through an iteration previously. By moving the construction of the
`range`s out of the `for`-loop each `range` is constructed once. Then
only the product of `range`s need be constructed on entering the
`for`-loop (as this dependent on how many `range`s are used). Also reuse
these constructed `range`s in the first `for`-loop as well.