Raise FutureWarning for index with rank above 1
If the `index` has a rank greater than 1 (e.g. a matrix), SciPy ndimage
measurements' functions often handle these in some fashion. Though they
don't always handle this case nor do they handle it consistently. We
have tried to mimic their behavior here, but it is sometimes a little
strange. So as to discourage users from making use of this behavior, we
raise a `FutureWarning` to point out that it is subject to change.