Simplify handling of default labels and index
Avoid performing extra computations on the default cases of `labels` and
`index` by simply filling them with their default values. Also most
SciPy documentation mentions that `labels` defaults to the whole image.
We have been restricting it to the non-zero portion, but that is
technically incorrect. Changing it to be the whole image hasn't had any
negative effects before, but we go ahead and do it now to match SciPy's
docs more accurately.