Freshen up API documentation
Try to fix up the API documentation to be more consistent with how the
Dask-based functions are written. Mainly this means pointing out things
are arrays. Though it also means streamlining documentation in terms of
how common default parameters are handled and how the results are
explained. Documentation is aimed to be clear, but also brief to avoid
redundancy or the potential of confusion.