Rewrite maximum_position to use argmax
Makes use of `labeled_comprehension` to rewrite `maximum_position`
around `argmax`. This way there is no need to compute `maximum` first to
find `maximum_position`. Also this removes a bunch of other steps that
`maximum_position` had added in previously to reuse `maximum` to get the
final result. The net result should be a cleaner implementation that is
easier to understand and maintain. Plus this ends up being ~33% faster
than the original implementation.