Rewrite minimum_position to use argmin
Makes use of `labeled_comprehension` to rewrite `minimum_position`
around `argmin`. This way there is no need to compute `minimum` first to
find `minimum_position`. Also this removes a bunch of other steps that
`minimum_position` had added in previously to reuse `minimum` to get the
final result. The net result should be a cleaner implementation that is
easier to understand and maintain. Plus this ends up being ~33% faster
than the original implementation.