Use an iterable range in core tests
Was using `range` on both Python 2 and Python 3. The problem being that
on Python 2 this allocates a `list` with the contents of everything
included by `range`'s bounds and step size. To fix this, simply use a
Python 2/3 compatibility trick to get an iterable range on Python 2.
Behavior on Python 3 is unchanged (excepting an alias).