Use labeled_comprehension in sum
Rewrites our `sum` implementation to make use of `labeled_comprehension`
for performing the computation. This should provide some more coverage
of `labeled_comprehension` and give us more confidence in using it in
other cases. Also as masks are used to select out the data of interest,
we are able to avoid performing computations on values that are not of
interest (i.e. not in the mask). As such we are able to achieve a not
insignificant speedup (~40%).