Create initial content (#1)
* Generate with dask-image/dask-image-cookiecutter

Use a cookiecutter template to generate all of the contents for
dask-image to start with. Will need some tweaking afterwards, but this
at least provides a good starting point.

* Wrap license text, add years, note authors

Still a BSD 3-Clause license. No actually license content changed. Did
wrap the text though as GitHub doesn't do that for us with text files.
Also updated the years and noted the copyright belongs to dask-image
developers.

* Update license metadata to include clauses

* Fix URLs to point to Dask org

The cookiecutter template was designed to template these with the
relevant GitHub username. Hence this ends up being incorrect and needed
to be fixed.

* Clean up conda environment name

* Move PR guidelines into PR template

* Strip some extra whitespace in Travis CI config