Refresh repo configuration
This adds support for AppVeyor for Windows and CircleCI for Linux. It
also converts Travis CI to macOS now. Testing environment files are
split apart based on OS and Python version to make them easier to
administer. Manages installation of Miniconda and creation of the
testing environment in each build using scripts. Should make it easy to
test locally if needed. Fast finishes CI builds on PRs when new ones are
pushed.