Support SciPy 1.0.0+ too (#37)
* Relax SciPy version constraint

Allow users to install with SciPy 1.0.0+.

* Use SciPy 1.0.0 on CIs

Start testing with SciPy 1.0.0 on CIs.

* Skip some tests that fail due to a SciPy 1.0.0 bug

Appears that SciPy 1.0.0+ has a bug that causes it to incorrectly
compute some Fourier filters when the data has double precision. So add
a check to skip these cases. We still test the filters for other
precision. So this doesn't block us from testing them as well.

* Switch back to SciPy 0.19.1 on CIs

As coverage drops when testing with SciPy 1.0.0+ due to some tests that
need to be skipped due to a SciPy bug, switch back to testing with SciPy
0.19.1. After all there appear to be no other significant differences
between the two versions of SciPy.