General cleanup (#39)
* Drop stub `dask_image.core` file

This file was generated by the cookiecutter template and is no longer
needed or used. So just go ahead and drop it.

* Fill out feature section of README

Lists some common functionality included in this package.

* Drop Credits section from the README

This was generated using Cookiecutter. Though based on feedback this was
deemed unpolished. So this drops it.

* Drop the Usage doc section

This was created by the cookiecutter template, but is not being used.
Instead users should read the API documentation for how to use different
functions. Hopefully this will be filled out in time with some better
examples.

* Improve index page title

* Convert `imread` into a subpackage

This shouldn't be strictly necessary. However this does allow `imread`
to match other subpackages in `dask-image`.

* Bump `pims` requirement for RTD docs

In the ReadTheDocs build, PIMS 0.3.3 has a bug revealed by a change in
six. However this was fixed in subsequent versions of PIMS. So just go
ahead and bump the PIMS requirement to 0.4.1 to match what we are using
in CI.

* Refresh `setup.py` requirements

Go ahead and refresh the requirements in `setup.py` to match those being
used in CIs. While the older versions likely work without issues, it is
easier to just require slightly newer (though still rather old) versions
of our dependencies generally. Should make it easier for packagers to
follow as well.

* Explicitly require `dask[array]`

We use Dask particularly with Dask Array support. For `pip` this means,
we should specify the extras we need as well. So require `dask[array]`
in `setup.py` to make sure these extra dependencies are pulled in too.