Consolidate README content in docs (#40)
* Move features to index page  [ci skip]

Based on feedback, move the list of features into the index page and
drop them from the README.

* Drop README from docs  [ci skip]

As most of the content is already in the docs, the README doesn't need
to be published in the docs. So drop it from the index and remove the
README page.

* Make "Indices and tables" a subheading  [ci skip]

Looked weird when this was a heading of the same size as the title.
Especially with other subheadings present. So this fixes that issue.