Fix test__norm_input_labels_index
The `ind_n` array is expected to be a 0-D array, but this check assumed
it was a 1-D singleton array, which is incorrect. Later versions of Dask
fixed `assert_eq` to better ensure that 0-D scalar arrays were not
compared equal to 1-D singleton arrays. So this fixes the test to behave
correctly.