Fix bug in test__norm_input_labels_index (#43)
* Fix test__norm_input_labels_index

The `ind_n` array is expected to be a 0-D array, but this check assumed
it was a 1-D singleton array, which is incorrect. Later versions of Dask
fixed `assert_eq` to better ensure that 0-D scalar arrays were not
compared equal to 1-D singleton arrays. So this fixes the test to behave
correctly.

* Test shapes in test__norm_input_labels_index

Make sure to explicitly test shapes of the normalized arrays in
`test__norm_input_labels_index`. Older versions of Dask would treat 1-D
scalar arrays as equivalent to 0-D arrays. So this check is important to
ensure that we are actually verifying `shape` and `ndim` are as
expected.

* Note test fix in history