Breakout common changes in fourier_gaussian (#45)
In all cases `sigma` is squared, the result is divided by `2`, and the
negative of the result taken. These can be factored out into a common
change to just `sigma` that is reused in all of the other computations.
Admittedly Dask is smart enough to capture the reused computation of
`sigma` squared, but it is not a symbolic algebra library. So it won't
be able to pick up on the other changes, but we can. This should
simplify the resulting graph and make it easy to reuse some common
changes to `sigma` before `tensordot` is computed over multiple arrays
and these changes applied.