Ensure `J` is negated first in `fourier_shift` (#46)
Just to ensure that `J` is negated before graph construction and it is
not rolled into graph construction, enforce the order of operations by
treating `-J` as a quantity by itself. This way we can be sure it is
already computed before graph construction begins and the scalar is
merely inserted into the graph as a single node.