Perform `2 * pi` first in `_get_ang_freq_grid` (#47)
Just to ensure this scalar multiplication occurs before graph
construction and is not added as operations in graph construction,
enforce the order of operations by treating `2 * pi` as a quantity
itself. This way we can be sure it is already computed before graph
construction begins and the scalar is merely inserted into the graph as
a single node.