Drop duplicate utility tests from `ndmorph` (#55)
These tests are remnants from when these utility functions were part of
`ndmorph`. However they are also part of `ndfilters`, which also has
tests for them. As these utility functions are now just used from
`ndfilters` in `ndmorph` and are tested in `ndfilters`, there is no need
to have the same tests in `ndmorph`. Hence this removes those tests,
which are duplicated. Should cutdown maintenance burden and speed up the
test suite a bit.