Rewrite `center_of_mass` to use `arange`
Instead of using the vendored `indices` function to multiply the image
with, rewrite `center_of_mass` to just use `arange` and rely on Dask to
handle the broadcasting when multiplying with the image. Should be a bit
more efficient and should avoid the need for a compatibility function
using `indices` in `center_of_mass`.