Rewrite `_ravel_shape_indices` without `indices`
Instead of relying on a vendored copy of `indices` to construct the
raveled shape indices, just use `arange` for each dimension and rely on
Dask to handle the broadcasting for us. Also handle any scalar
multiplication that would occur within the arguments to `arange`. This
creates a much simpler Dask graph and avoids relying on a vendored copy
of `indices`.