Drop note in `_ravel_shape_indices`
As we now are leveraging just `arange` to handle construction of the
indices in `_ravel_shape_indices`, this note no longer applies. Plus we
are better able to handle chunking with what we are doing in this code
than would happen with `reshape` naively. So this really isn't a
workaround any more, but an efficient solution to the problem.