Bump Dask requirement to 0.16.1+
The Dask 0.13.0 requirement was in place before as that was a pretty
stable and somewhat old Dask version at the time and Python 3.4 packages
existed for it on conda-forge. Also it was believed this should make it
easier for other projects to use this Dask-based image processing code.

However many changes were pushed upstream to Dask when this was being
developed, which were largely vendored or worked around in this library
through compatibility code. In some cases the upstream implementations
have seen notable performance improvements. In other cases,
functionality (like Dask Arrays of unknown shape) have just become
easier to work with generally with fewer hacks. Of course there is also
the maintenance that this extra code entails, which is not really worth
the effort given upstream often has superior implementations.Generally
having this compatibility code is becoming a bit of a hindrance.

Also Python 3.4 is not being supported by Dask further or many of the
projects downstream of it. Further some of these projects already
require substantially newer versions of Dask to function. As we may have
similar implementations of things that differ as noted above, this can
be costly in terms of interoperability. So supporting such old versions
of Dask is really not helpful for others wanting to use this library.

Given all of this, it makes sense to bump the Dask minimum version for
better performance, a simpler code base, and nicer interoperability with
other libraries. The minimum version of Dask that would help us
eliminate all compatibility code would be 0.16.1. While a newer version
could be used for other new features, none of these are needed yet. As
0.16.1 came out early this year, it has been around for a while. It is
also the last patch release of the 0.16 series. Plus 0.16.0 introduces
the Dask Collections interface to the public API, which makes it easier
to construct custom graphs should they be needed. Thus go ahead and bump
the minimum Dask version to 0.16.1.