Drop all `_compat` modules
As the `_compat` modules were added to provide Dask Array functionality
not yet included in the Dask version we were using, they are now
unneeded given our minimum version of Dask has been bumped up to a
version that includes all of this functionality. So just remove the
`_compat` modules from the codebase.