Use `meshgrid` in `_get_freq_grid`
Instead of rolling our own version of `meshgrid` in `_get_freq_grid` to
construct the frequency grid, simply pass the results from `fftfreq`
onto `meshgrid`. This drastically cuts down on the amount of code here
and benefits from all of the optimizations upstream has implemented in
`meshgrid`, which we did not have here with our `repeat`-based
implementation.