Use Dask Array's `where` directly
We had wrapped Dask Array's `where` to perform a quick optimization if
the condition was a simple scalar (e.g. `True` or `False`). In this
special case, we merely selected the correct result and returned it
instead of performing a lengthy computation, whose result was already
known during graph computation. This optimization has long since made
its way into Dask Array. So there is no value in us holding onto this
wrapper code any longer. Hence it is dropped.