Drop tests for the `ndmorph._ops` module
The only test of `ndmorph._ops` was for the `_where` function, which we
no longer need. The other functionality in there is a wrapper that is
already effectively tested by the construction of all `ndmorph`
functions and their tests. Having an explicit test of the wrapper makes
no sense since it already occurs through the functions it is used to
make. Given this, just drop this test module since it is no longer
needed.